+++ tests/Frontend/Router.spec.js
@@ -5,6 +5,7 @@ import { createTestingPinia } from '@pinia/testing';
 import { useLoadingStore } from '../../resources/js/stores/loading';
 import { useAuthStore } from '../../resources/js/stores/auth';
 import { mockAxios } from './helper';
+import { it } from 'vitest';
 
 const accessPermittedRolesView = routes.filter(route => route.path === '/settings')[0]
   .children.filter(route => route.name === 'settings.roles.view')[0].meta.accessPermitted;
@@ -37,6 +38,26 @@ describe('Router', () => {
   });
 
   describe('beforeEachRoute', () => {
+    it('beforeEachRoute sets the application title', async () => {
+      const router = {};
+
+      createTestingPinia({ initialState: { settings: { settings: { name: 'Appname' } } }, stubActions: false });
+      const loading = useLoadingStore();
+      loading.initialized = true;
+
+      const to = {
+        matched: [{ path: '/', meta: {} }]
+      };
+
+      await new Promise((resolve) => {
+        beforeEachRoute(router, to, undefined, () => {
+          resolve();
+        });
+      });
+
+      expect(document.title).toBe('Appname');
+    });
+
     it('beforeEachRoute calls next if there is no permission checks or required authentication', async () => {
       const router = {};